Remove DefaultCirclePiece
